fix bug with sorting by custom_asset_type for custom_assets
